Add new friends service and move to services dir
