+++ src/app/services/friends.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from "@angular/core";
+
+@Injectable({
+	providedIn: "root",
+})
+export class FriendsService {
+	constructor() {}
+}
